requirements.txt
matplotlib==3.10.7
numpy==2.2.6
open_clip_torch==3.2.0
opencv-python==4.11.0.86
pandas==2.3.3
pillow==12.0.0
scikit-learn==1.7.2
scipy==1.15.3
tensorboard==2.19.0
torch==2.9.1
torchmetrics==1.6.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3